move bsd_utils.h to a better place
