+++ src/CMakeLists.txt
@@ -36,7 +36,6 @@ set(headers
     include/rasta/rastaredundancy.h
     include/rasta/rastautil.h
     include/rasta/rmemory.h
-    include/rasta/bsd_utils.h
     include/rasta/rastablake2.h
     include/rasta/rastasiphash24.h
     include/rasta/rastahashing.h
@@ -47,6 +46,7 @@ set(sources
     c/logging.c
     c/rastafactory.c
     c/transport/bsd_utils.c
+    c/transport/bsd_utils.h
     c/transport/diagnostics.c
     c/transport/diagnostics.h
     c/transport/events.c
