+++ src/c/transport/tcp_base.c
@@ -1,7 +1,9 @@
-#include "tcp.h"
-#include <rasta/bsd_utils.h>
-#include <rasta/rmemory.h>
 #include <stdlib.h>
+
+#include <rasta/rmemory.h>
+
+#include "bsd_utils.h"
+#include "tcp.h"
 
 void transport_create_socket(struct rasta_handle *h, rasta_transport_socket *socket, int id, const rasta_config_tls *tls_config) {
     // init socket
